Fix minor typo, name imrovment and move travelGuaranteeFieldValidator into the dir special-field-validators to make the distinction between validation of common form fields and special cases more clear.
