+++ src/page-modules/contact/validation/index.ts
@@ -1,3 +1,3 @@
 export { default as commonFieldValidator } from './commonFieldValidator';
-export { default as travelGuaranteeFieldValidator } from './travelGuaranteeFieldValidator';
+export { default as travelGuaranteeFieldValidator } from './special-field-validators/travelGuaranteeFieldValidator';
 export { addErrorMessage, type InputErrorMessages } from './utils';
